Move search results props to types file
