+++ src/types/types.d.ts
@@ -147,3 +147,9 @@ export type OpenSpaceTimeState = {
   prevStep?: number;
   deltaTimeSteps?: number[];
 };
+
+export interface SearchResultsProps<T> {
+  data: T[];
+  renderElement: (data: T, i: number) => React.ReactNode;
+  matcherFunc: (data: T, searchString: string) => boolean;
+}
